+++ ahbottomnavigation/src/main/java/com/aurelhubert/ahbottomnavigation/AHBottomNavigation.java
@@ -1,8 +1,11 @@
 package com.aurelhubert.ahbottomnavigation;
 
 import android.animation.Animator;
+import android.annotation.SuppressLint;
+import android.annotation.TargetApi;
 import android.content.Context;
 import android.content.res.Resources;
+import android.content.res.TypedArray;
 import android.graphics.Color;
 import android.graphics.Typeface;
 import android.graphics.drawable.Drawable;
@@ -11,18 +14,23 @@ import android.os.Bundle;
 import android.os.Parcelable;
 import android.support.annotation.ColorInt;
 import android.support.annotation.ColorRes;
+import android.support.annotation.DrawableRes;
 import android.support.design.widget.CoordinatorLayout;
+import android.support.design.widget.FloatingActionButton;
 import android.support.v4.content.ContextCompat;
 import android.support.v4.view.ViewCompat;
 import android.support.v4.view.animation.LinearOutSlowInInterpolator;
 import android.util.AttributeSet;
+import android.util.DisplayMetrics;
 import android.util.Log;
 import android.util.TypedValue;
+import android.view.Display;
 import android.view.Gravity;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewAnimationUtils;
 import android.view.ViewGroup;
+import android.view.WindowManager;
 import android.view.animation.AccelerateInterpolator;
 import android.view.animation.OvershootInterpolator;
 import android.widget.FrameLayout;
@@ -31,9 +39,12 @@ import android.widget.LinearLayout;
 import android.widget.TextView;
 
 import com.squareup.picasso.Picasso;
+import com.aurelhubert.ahbottomnavigation.notification.AHNotification;
+import com.aurelhubert.ahbottomnavigation.notification.AHNotificationHelper;
 
 import java.util.ArrayList;
 import java.util.List;
+import java.util.Locale;
 
 /**
  * AHBottomNavigationLayout
@@ -45,13 +56,22 @@ public class AHBottomNavigation extends FrameLayout {
 	public static final int CURRENT_ITEM_NONE = -1;
 	public static final int UPDATE_ALL_NOTIFICATIONS = -1;
 
+	// Title state
+	public enum TitleState {
+		SHOW_WHEN_ACTIVE,
+		ALWAYS_SHOW,
+		ALWAYS_HIDE
+	}
+
 	// Static
 	private static String TAG = "AHBottomNavigation";
+    private static final String EXCEPTION_INDEX_OUT_OF_BOUNDS = "The position (%d) is out of bounds of the items (%d elements)";
 	private static final int MIN_ITEMS = 3;
 	private static final int MAX_ITEMS = 5;
 
 	// Listener
 	private OnTabSelectedListener tabSelectedListener;
+	private OnNavigationPositionListener navigationPositionListener;
 
 	// Variables
 	private Context context;
@@ -59,30 +79,37 @@ public class AHBottomNavigation extends FrameLayout {
 	private ArrayList<AHBottomNavigationItem> items = new ArrayList<>();
 	private ArrayList<View> views = new ArrayList<>();
 	private AHBottomNavigationBehavior<AHBottomNavigation> bottomNavigationBehavior;
+	private LinearLayout linearLayoutContainer;
 	private View backgroundColorView;
 	private Animator circleRevealAnim;
 	private boolean colored = false;
-	private String[] notifications = {"", "", "", "", ""};
+	private boolean selectedBackgroundVisible = false;
+	private boolean translucentNavigationEnabled;
+	private boolean isHidden = false;
+	private List<AHNotification> notifications = AHNotification.generateEmptyList(MAX_ITEMS);
 	private boolean isBehaviorTranslationSet = false;
 	private int currentItem = 0;
 	private int currentColor = 0;
 	private boolean behaviorTranslationEnabled = true;
 	private boolean needHideBottomNavigation = false;
 	private boolean hideBottomNavigationWithAnimation = false;
+	private boolean soundEffectsEnabled = true;
 
 	// Variables (Styles)
 	private Typeface titleTypeface;
 	private int defaultBackgroundColor = Color.WHITE;
-	//private int accentColor = Color.WHITE;
-	//private int inactiveColor = Color.WHITE;
-	private @ColorInt int itemActiveColor, itemInactiveColor;
-	private @ColorInt int titleColorActive, titleColorInactive;
-	private @ColorInt int coloredTitleColorActive, coloredTitleColorInactive;
+	private int defaultBackgroundResource = 0;
+	private @ColorInt int itemActiveColor;
+	private @ColorInt int itemInactiveColor;
+	private @ColorInt int titleColorActive;
+	private @ColorInt int titleColorInactive;
+	private @ColorInt int coloredTitleColorActive;
+	private @ColorInt int coloredTitleColorInactive;
 	private float titleActiveTextSize, titleInactiveTextSize;
-	private int bottomNavigationHeight;
+	private int bottomNavigationHeight, navigationBarHeight = 0;
 	private float selectedItemWidth, notSelectedItemWidth;
 	private boolean forceTint = false;
-	private boolean forceTitlesDisplay = false;
+	private TitleState titleState = TitleState.SHOW_WHEN_ACTIVE;
 
 	// Notifications
 	private @ColorInt int notificationTextColor;
@@ -90,23 +117,30 @@ public class AHBottomNavigation extends FrameLayout {
 	private Drawable notificationBackgroundDrawable;
 	private Typeface notificationTypeface;
 	private int notificationActiveMarginLeft, notificationInactiveMarginLeft;
+	private int notificationActiveMarginTop, notificationInactiveMarginTop;
 
 	/**
 	 * Constructors
 	 */
 	public AHBottomNavigation(Context context) {
 		super(context);
-		init(context);
+		init(context, null);
 	}
 
 	public AHBottomNavigation(Context context, AttributeSet attrs) {
 		super(context, attrs);
-		init(context);
+		init(context, attrs);
 	}
 
 	public AHBottomNavigation(Context context, AttributeSet attrs, int defStyleAttr) {
 		super(context, attrs, defStyleAttr);
-		init(context);
+		init(context, attrs);
+	}
+
+	@Override
+	public void setSoundEffectsEnabled(final boolean soundEffectsEnabled) {
+		super.setSoundEffectsEnabled(soundEffectsEnabled);
+		this.soundEffectsEnabled = soundEffectsEnabled;
 	}
 
 	@Override
@@ -130,7 +164,7 @@ public class AHBottomNavigation extends FrameLayout {
 		Bundle bundle = new Bundle();
 		bundle.putParcelable("superState", super.onSaveInstanceState());
 		bundle.putInt("current_item", currentItem);
-		bundle.putStringArray("notifications", notifications);
+        bundle.putParcelableArrayList("notifications", new ArrayList<> (notifications));
 		return bundle;
 	}
 
@@ -139,7 +173,7 @@ public class AHBottomNavigation extends FrameLayout {
 		if (state instanceof Bundle) {
 			Bundle bundle = (Bundle) state;
 			currentItem = bundle.getInt("current_item");
-			notifications = bundle.getStringArray("notifications");
+            notifications = bundle.getParcelableArrayList("notifications");
 			state = bundle.getParcelable("superState");
 		}
 		super.onRestoreInstanceState(state);
@@ -154,9 +188,19 @@ public class AHBottomNavigation extends FrameLayout {
 	 *
 	 * @param context
 	 */
-	private void init(Context context) {
+	private void init(Context context, AttributeSet attrs) {
 		this.context = context;
 		resources = this.context.getResources();
+
+		if (attrs != null) {
+			TypedArray ta = context.obtainStyledAttributes(attrs, R.styleable.AHBottomNavigationBehavior_Params, 0, 0);
+			try {
+				selectedBackgroundVisible = ta.getBoolean(R.styleable.AHBottomNavigationBehavior_Params_selectedBackgroundVisible, false);
+				translucentNavigationEnabled = ta.getBoolean(R.styleable.AHBottomNavigationBehavior_Params_translucentNavigationEnabled, false);
+			} finally {
+				ta.recycle();
+			}
+		}
 
 		notificationTextColor = ContextCompat.getColor(context, android.R.color.white);
 		bottomNavigationHeight = (int) resources.getDimension(R.dimen.bottom_navigation_height);
@@ -174,6 +218,8 @@ public class AHBottomNavigation extends FrameLayout {
 		// Notifications
 		notificationActiveMarginLeft = (int) resources.getDimension(R.dimen.bottom_navigation_notification_margin_left_active);
 		notificationInactiveMarginLeft = (int) resources.getDimension(R.dimen.bottom_navigation_notification_margin_left);
+		notificationActiveMarginTop = (int) resources.getDimension(R.dimen.bottom_navigation_notification_margin_top_active);
+		notificationInactiveMarginTop = (int) resources.getDimension(R.dimen.bottom_navigation_notification_margin_top);
 
 		ViewCompat.setElevation(this, resources.getDimension(R.dimen.bottom_navigation_elevation));
 		setClipToPadding(false);
@@ -200,21 +246,23 @@ public class AHBottomNavigation extends FrameLayout {
 		backgroundColorView = new View(context);
 		if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.LOLLIPOP) {
 			LayoutParams backgroundLayoutParams = new LayoutParams(
-					ViewGroup.LayoutParams.MATCH_PARENT, layoutHeight);
+					ViewGroup.LayoutParams.MATCH_PARENT, calculateHeight(layoutHeight));
 			addView(backgroundColorView, backgroundLayoutParams);
+			bottomNavigationHeight = layoutHeight;
 		}
 
-		LinearLayout linearLayout = new LinearLayout(context);
-		linearLayout.setOrientation(LinearLayout.HORIZONTAL);
-		linearLayout.setGravity(Gravity.CENTER);
+		linearLayoutContainer = new LinearLayout(context);
+		linearLayoutContainer.setOrientation(LinearLayout.HORIZONTAL);
+		linearLayoutContainer.setGravity(Gravity.CENTER);
 
 		LayoutParams layoutParams = new LayoutParams(ViewGroup.LayoutParams.MATCH_PARENT, layoutHeight);
-		addView(linearLayout, layoutParams);
+		addView(linearLayoutContainer, layoutParams);
 
-		if (items.size() == MIN_ITEMS || forceTitlesDisplay) {
-			createClassicItems(linearLayout);
+		if (titleState != TitleState.ALWAYS_HIDE &&
+				(items.size() == MIN_ITEMS || titleState == TitleState.ALWAYS_SHOW)) {
+			createClassicItems(linearLayoutContainer);
 		} else {
-			createSmallItems(linearLayout);
+			createSmallItems(linearLayoutContainer);
 		}
 
 		// Force a request layout after all the items have been created
@@ -225,6 +273,65 @@ public class AHBottomNavigation extends FrameLayout {
 			}
 		});
 	}
+
+	@SuppressLint("NewApi")
+	@TargetApi(Build.VERSION_CODES.LOLLIPOP)
+	private int calculateHeight(int layoutHeight) {
+		if(!translucentNavigationEnabled) return layoutHeight;
+
+		int resourceId = getResources().getIdentifier("navigation_bar_height", "dimen", "android");
+		if (resourceId > 0) {
+			navigationBarHeight = resources.getDimensionPixelSize(resourceId);
+		}
+
+		int[] attrs = {android.R.attr.fitsSystemWindows, android.R.attr.windowTranslucentNavigation};
+		TypedArray typedValue = getContext().getTheme().obtainStyledAttributes(attrs);
+
+		@SuppressWarnings("ResourceType")
+		boolean fitWindow = typedValue.getBoolean(0, false);
+
+		@SuppressWarnings("ResourceType")
+		boolean translucentNavigation = typedValue.getBoolean(1, true);
+
+		if(hasImmersive() /*&& !fitWindow*/ && translucentNavigation) {
+			layoutHeight += navigationBarHeight;
+		}
+
+		typedValue.recycle();
+
+		return layoutHeight;
+	}
+
+	@SuppressLint("NewApi")
+	@TargetApi(Build.VERSION_CODES.LOLLIPOP)
+	public boolean hasImmersive() {
+		Display d = ((WindowManager)getContext().getSystemService(Context.WINDOW_SERVICE)).getDefaultDisplay();
+
+		DisplayMetrics realDisplayMetrics = new DisplayMetrics();
+		d.getRealMetrics(realDisplayMetrics);
+
+		int realHeight = realDisplayMetrics.heightPixels;
+		int realWidth = realDisplayMetrics.widthPixels;
+
+		DisplayMetrics displayMetrics = new DisplayMetrics();
+		d.getMetrics(displayMetrics);
+
+		int displayHeight = displayMetrics.heightPixels;
+		int displayWidth = displayMetrics.widthPixels;
+
+		return (realWidth > displayWidth) || (realHeight > displayHeight);
+	}
+
+	// updated
+
+    /**
+     * Check if items must be classic
+     *
+     * @return true if classic (icon + title)
+     */
+    private boolean isClassic() {
+        return titleState == TitleState.ALWAYS_SHOW || items.size() <= MIN_ITEMS && titleState != TitleState.ALWAYS_SHOW;
+    }
 
 	/**
 	 * Create classic items (only 3 items in the bottom navigation)
@@ -239,7 +346,7 @@ public class AHBottomNavigation extends FrameLayout {
 		float minWidth = resources.getDimension(R.dimen.bottom_navigation_min_width);
 		float maxWidth = resources.getDimension(R.dimen.bottom_navigation_max_width);
 
-		if (forceTitlesDisplay && items.size() > MIN_ITEMS) {
+		if (titleState == TitleState.ALWAYS_SHOW && items.size() > MIN_ITEMS) {
 			minWidth = resources.getDimension(R.dimen.bottom_navigation_small_inactive_min_width);
 			maxWidth = resources.getDimension(R.dimen.bottom_navigation_small_inactive_max_width);
 		}
@@ -263,13 +370,13 @@ public class AHBottomNavigation extends FrameLayout {
 		if (titleActiveTextSize != 0 && titleInactiveTextSize != 0) {
 			activeSize = titleActiveTextSize;
 			inactiveSize = titleInactiveTextSize;
-		} else if (forceTitlesDisplay && items.size() > MIN_ITEMS) {
+		} else if (titleState == TitleState.ALWAYS_SHOW && items.size() > MIN_ITEMS) {
 			activeSize = resources.getDimension(R.dimen.bottom_navigation_text_size_forced_active);
 			inactiveSize = resources.getDimension(R.dimen.bottom_navigation_text_size_forced_inactive);
 		}
 
 		for (int i = 0; i < items.size(); i++) {
-
+			final boolean current = currentItem == i;
 			final int itemIndex = i;
 			AHBottomNavigationItem item = items.get(itemIndex);
 
@@ -287,11 +394,14 @@ public class AHBottomNavigation extends FrameLayout {
 				title.setTypeface(titleTypeface);
 			}
 
-			if (forceTitlesDisplay && items.size() > MIN_ITEMS) {
+			if (titleState == TitleState.ALWAYS_SHOW && items.size() > MIN_ITEMS) {
 				container.setPadding(0, container.getPaddingTop(), 0, container.getPaddingBottom());
 			}
 
-			if (i == currentItem) {
+			if (current) {
+				if (selectedBackgroundVisible) {
+					view.setSelected(true);
+				}
 				icon.setSelected(true);
 				// Update margins (icon & notification)
 				if (view.getLayoutParams() instanceof ViewGroup.MarginLayoutParams) {
@@ -314,24 +424,29 @@ public class AHBottomNavigation extends FrameLayout {
 			}
 
 			if (colored) {
-				if (i == currentItem) {
+				if (current) {
 					setBackgroundColor(item.getColor(context));
 					currentColor = item.getColor(context);
 				}
 			} else {
-				setBackgroundColor(defaultBackgroundColor);
+				if (defaultBackgroundResource != 0) {
+					setBackgroundResource(defaultBackgroundResource);
+				} else {
+					setBackgroundColor(defaultBackgroundColor);
+				}
 			}
 
 			icon.setImageDrawable(AHHelper.getTintDrawable(items.get(i).getDrawable(context),
-					currentItem == i ? itemActiveColor : itemInactiveColor, forceTint));
-			title.setTextColor(currentItem == i ? itemActiveColor :itemInactiveColor);
-			title.setTextSize(TypedValue.COMPLEX_UNIT_PX, currentItem == i ? activeSize : inactiveSize);
+					current ? itemActiveColor : itemInactiveColor, forceTint));
+			title.setTextColor(current ? itemActiveColor : itemInactiveColor);
+			title.setTextSize(TypedValue.COMPLEX_UNIT_PX, current ? activeSize : inactiveSize);
 			view.setOnClickListener(new OnClickListener() {
 				@Override
 				public void onClick(View v) {
 					updateItems(itemIndex, true);
 				}
 			});
+			view.setSoundEffectsEnabled(soundEffectsEnabled);
 
 			LayoutParams params = new LayoutParams((int) itemWidth, (int) height);
 			linearLayout.addView(view, params);
@@ -385,8 +500,12 @@ public class AHBottomNavigation extends FrameLayout {
 			TextView title = (TextView) view.findViewById(R.id.bottom_navigation_small_item_title);
 			TextView notification = (TextView) view.findViewById(R.id.bottom_navigation_notification);
 
-            Picasso.with(context).load(item.getDrawable()).resize(64, 64).into(icon);
-			title.setText(item.getTitle(context));
+			Picasso.with(context).load(item.getDrawable()).resize(64, 64).into(icon);
+			//icon.setImageDrawable(item.getDrawable(context));
+
+			if (titleState != TitleState.ALWAYS_HIDE) {
+				title.setText(item.getTitle(context));
+			}
 
 			if (titleActiveTextSize != 0) {
 				title.setTextSize(TypedValue.COMPLEX_UNIT_PX, titleActiveTextSize);
@@ -397,24 +516,30 @@ public class AHBottomNavigation extends FrameLayout {
 			}
 
 			if (i == currentItem) {
+				if (selectedBackgroundVisible) {
+					view.setSelected(true);
+				}
 				icon.setSelected(true);
 				// Update margins (icon & notification)
-				if (view.getLayoutParams() instanceof ViewGroup.MarginLayoutParams) {
-					ViewGroup.MarginLayoutParams p = (ViewGroup.MarginLayoutParams) icon.getLayoutParams();
-					p.setMargins(p.leftMargin, activeMarginTop, p.rightMargin, p.bottomMargin);
 
-					ViewGroup.MarginLayoutParams paramsNotification = (ViewGroup.MarginLayoutParams)
-							notification.getLayoutParams();
-					paramsNotification.setMargins(notificationActiveMarginLeft, paramsNotification.topMargin,
-							paramsNotification.rightMargin, paramsNotification.bottomMargin);
+				if (titleState != TitleState.ALWAYS_HIDE) {
+					if (view.getLayoutParams() instanceof ViewGroup.MarginLayoutParams) {
+						ViewGroup.MarginLayoutParams p = (ViewGroup.MarginLayoutParams) icon.getLayoutParams();
+						p.setMargins(p.leftMargin, activeMarginTop, p.rightMargin, p.bottomMargin);
 
-					view.requestLayout();
+						ViewGroup.MarginLayoutParams paramsNotification = (ViewGroup.MarginLayoutParams)
+								notification.getLayoutParams();
+						paramsNotification.setMargins(notificationActiveMarginLeft, notificationActiveMarginTop,
+								paramsNotification.rightMargin, paramsNotification.bottomMargin);
+
+						view.requestLayout();
+					}
 				}
 			} else {
 				icon.setSelected(false);
 				ViewGroup.MarginLayoutParams paramsNotification = (ViewGroup.MarginLayoutParams)
 						notification.getLayoutParams();
-				paramsNotification.setMargins(notificationInactiveMarginLeft, paramsNotification.topMargin,
+				paramsNotification.setMargins(notificationInactiveMarginLeft, notificationInactiveMarginTop,
 						paramsNotification.rightMargin, paramsNotification.bottomMargin);
 			}
 
@@ -424,7 +549,11 @@ public class AHBottomNavigation extends FrameLayout {
 					currentColor = item.getColor(context);
 				}
 			} else {
-				setBackgroundColor(defaultBackgroundColor);
+				if (defaultBackgroundResource != 0) {
+					setBackgroundResource(defaultBackgroundResource);
+				} else {
+					setBackgroundColor(defaultBackgroundColor);
+				}
 			}
 
 			icon.setImageDrawable(AHHelper.getTintDrawable(items.get(i).getDrawable(context),
@@ -437,9 +566,16 @@ public class AHBottomNavigation extends FrameLayout {
 					updateSmallItems(itemIndex, true);
 				}
 			});
+			view.setSoundEffectsEnabled(soundEffectsEnabled);
 
-			LayoutParams params = new LayoutParams(i == currentItem ? (int) selectedItemWidth :
-					(int) itemWidth, (int) height);
+			int width = i == currentItem ? (int) selectedItemWidth :
+					(int) itemWidth;
+
+			if (titleState == TitleState.ALWAYS_HIDE) {
+				width = (int) (itemWidth * 1.16);
+			}
+
+			LayoutParams params = new LayoutParams(width, (int) height);
 			linearLayout.addView(view, params);
 			views.add(view);
 		}
@@ -463,6 +599,11 @@ public class AHBottomNavigation extends FrameLayout {
 			return;
 		}
 
+		if (tabSelectedListener != null && useCallback) {
+			boolean selectionAllowed = tabSelectedListener.onTabSelected(itemIndex, false);
+			if (!selectionAllowed) return;
+		}
+
 		int activeMarginTop = (int) resources.getDimension(R.dimen.bottom_navigation_margin_top_active);
 		int inactiveMarginTop = (int) resources.getDimension(R.dimen.bottom_navigation_margin_top_inactive);
 		float activeSize = resources.getDimension(R.dimen.bottom_navigation_text_size_active);
@@ -471,18 +612,23 @@ public class AHBottomNavigation extends FrameLayout {
 		if (titleActiveTextSize != 0 && titleInactiveTextSize != 0) {
 			activeSize = titleActiveTextSize;
 			inactiveSize = titleInactiveTextSize;
-		} else if (forceTitlesDisplay && items.size() > MIN_ITEMS) {
+		} else if (titleState == TitleState.ALWAYS_SHOW && items.size() > MIN_ITEMS) {
 			activeSize = resources.getDimension(R.dimen.bottom_navigation_text_size_forced_active);
 			inactiveSize = resources.getDimension(R.dimen.bottom_navigation_text_size_forced_inactive);
 		}
 
 		for (int i = 0; i < views.size(); i++) {
 
+			final View view = views.get(i);
+			if (selectedBackgroundVisible) {
+				view.setSelected(i == itemIndex);
+			}
+
 			if (i == itemIndex) {
 
-				final TextView title = (TextView) views.get(itemIndex).findViewById(R.id.bottom_navigation_item_title);
-				final ImageView icon = (ImageView) views.get(itemIndex).findViewById(R.id.bottom_navigation_item_icon);
-				final TextView notification = (TextView) views.get(itemIndex).findViewById(R.id.bottom_navigation_notification);
+				final TextView title = (TextView) view.findViewById(R.id.bottom_navigation_item_title);
+				final ImageView icon = (ImageView) view.findViewById(R.id.bottom_navigation_item_icon);
+				final TextView notification = (TextView) view.findViewById(R.id.bottom_navigation_notification);
 
 				icon.setSelected(true);
 				AHHelper.updateTopMargin(icon, inactiveMarginTop, activeMarginTop);
@@ -495,8 +641,8 @@ public class AHBottomNavigation extends FrameLayout {
 				if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.LOLLIPOP && colored) {
 
 					int finalRadius = Math.max(getWidth(), getHeight());
-					int cx = (int) views.get(itemIndex).getX() + views.get(itemIndex).getWidth() / 2;
-					int cy = views.get(itemIndex).getHeight() / 2;
+					int cx = (int) view.getX() + view.getWidth() / 2;
+					int cy = view.getHeight() / 2;
 
 					if (circleRevealAnim != null && circleRevealAnim.isRunning()) {
 						circleRevealAnim.cancel();
@@ -531,15 +677,19 @@ public class AHBottomNavigation extends FrameLayout {
 					AHHelper.updateViewBackgroundColor(this, currentColor,
 							items.get(itemIndex).getColor(context));
 				} else {
-					setBackgroundColor(defaultBackgroundColor);
+					if (defaultBackgroundResource != 0) {
+						setBackgroundResource(defaultBackgroundResource);
+					} else {
+						setBackgroundColor(defaultBackgroundColor);
+					}
 					backgroundColorView.setBackgroundColor(Color.TRANSPARENT);
 				}
 
 			} else if (i == currentItem) {
 
-				final TextView title = (TextView) views.get(currentItem).findViewById(R.id.bottom_navigation_item_title);
-				final ImageView icon = (ImageView) views.get(currentItem).findViewById(R.id.bottom_navigation_item_icon);
-				final TextView notification = (TextView) views.get(currentItem).findViewById(R.id.bottom_navigation_notification);
+				final TextView title = (TextView) view.findViewById(R.id.bottom_navigation_item_title);
+				final ImageView icon = (ImageView) view.findViewById(R.id.bottom_navigation_item_icon);
+				final TextView notification = (TextView) view.findViewById(R.id.bottom_navigation_notification);
 
 				icon.setSelected(false);
 				AHHelper.updateTopMargin(icon, activeMarginTop, inactiveMarginTop);
@@ -555,12 +705,12 @@ public class AHBottomNavigation extends FrameLayout {
 		if (currentItem > 0 && currentItem < items.size()) {
 			currentColor = items.get(currentItem).getColor(context);
 		} else if (currentItem == CURRENT_ITEM_NONE) {
-			setBackgroundColor(defaultBackgroundColor);
+			if (defaultBackgroundResource != 0) {
+				setBackgroundResource(defaultBackgroundResource);
+			} else {
+				setBackgroundColor(defaultBackgroundColor);
+			}
 			backgroundColorView.setBackgroundColor(Color.TRANSPARENT);
-		}
-
-		if (tabSelectedListener != null && useCallback) {
-			tabSelectedListener.onTabSelected(itemIndex, false);
 		}
 	}
 
@@ -579,24 +729,39 @@ public class AHBottomNavigation extends FrameLayout {
 			return;
 		}
 
+		if (tabSelectedListener != null && useCallback) {
+			boolean selectionAllowed = tabSelectedListener.onTabSelected(itemIndex, false);
+			if (!selectionAllowed) return;
+		}
+
 		int activeMarginTop = (int) resources.getDimension(R.dimen.bottom_navigation_small_margin_top_active);
 		int inactiveMargin = (int) resources.getDimension(R.dimen.bottom_navigation_small_margin_top);
 
 		for (int i = 0; i < views.size(); i++) {
 
+			final View view = views.get(i);
+			if (selectedBackgroundVisible) {
+				view.setSelected(i == itemIndex);
+			}
+
 			if (i == itemIndex) {
 
-				final FrameLayout container = (FrameLayout) views.get(itemIndex).findViewById(R.id.bottom_navigation_small_container);
-				final TextView title = (TextView) views.get(itemIndex).findViewById(R.id.bottom_navigation_small_item_title);
-				final ImageView icon = (ImageView) views.get(itemIndex).findViewById(R.id.bottom_navigation_small_item_icon);
-				final TextView notification = (TextView) views.get(itemIndex).findViewById(R.id.bottom_navigation_notification);
+				final FrameLayout container = (FrameLayout) view.findViewById(R.id.bottom_navigation_small_container);
+				final TextView title = (TextView) view.findViewById(R.id.bottom_navigation_small_item_title);
+				final ImageView icon = (ImageView) view.findViewById(R.id.bottom_navigation_small_item_icon);
+				final TextView notification = (TextView) view.findViewById(R.id.bottom_navigation_notification);
 
 				icon.setSelected(true);
-				AHHelper.updateTopMargin(icon, inactiveMargin, activeMarginTop);
-				AHHelper.updateLeftMargin(notification, notificationInactiveMarginLeft, notificationActiveMarginLeft);
-				AHHelper.updateTextColor(title, itemInactiveColor, itemActiveColor);
+
+				if (titleState != TitleState.ALWAYS_HIDE) {
+					AHHelper.updateTopMargin(icon, inactiveMargin, activeMarginTop);
+					AHHelper.updateLeftMargin(notification, notificationInactiveMarginLeft, notificationActiveMarginLeft);
+					AHHelper.updateTopMargin(notification, notificationInactiveMarginTop, notificationActiveMarginTop);
+					AHHelper.updateTextColor(title, itemInactiveColor, itemActiveColor);
+					AHHelper.updateWidth(container, notSelectedItemWidth, selectedItemWidth);
+				}
+
 				AHHelper.updateAlpha(title, 0, 1);
-				AHHelper.updateWidth(container, notSelectedItemWidth, selectedItemWidth);
 				AHHelper.updateDrawableColor(context, items.get(itemIndex).getDrawable(context), icon,
 						itemInactiveColor, itemActiveColor, forceTint);
 
@@ -638,26 +803,34 @@ public class AHBottomNavigation extends FrameLayout {
 					AHHelper.updateViewBackgroundColor(this, currentColor,
 							items.get(itemIndex).getColor(context));
 				} else {
-					setBackgroundColor(defaultBackgroundColor);
+					if (defaultBackgroundResource != 0) {
+						setBackgroundResource(defaultBackgroundResource);
+					} else {
+						setBackgroundColor(defaultBackgroundColor);
+					}
 					backgroundColorView.setBackgroundColor(Color.TRANSPARENT);
 				}
 
 			} else if (i == currentItem) {
 
-				final View container = views.get(currentItem).findViewById(R.id.bottom_navigation_small_container);
-				final TextView title = (TextView) views.get(currentItem).findViewById(R.id.bottom_navigation_small_item_title);
-				final ImageView icon = (ImageView) views.get(currentItem).findViewById(R.id.bottom_navigation_small_item_icon);
-				final TextView notification = (TextView) views.get(currentItem).findViewById(R.id.bottom_navigation_notification);
+				final View container = view.findViewById(R.id.bottom_navigation_small_container);
+				final TextView title = (TextView) view.findViewById(R.id.bottom_navigation_small_item_title);
+				final ImageView icon = (ImageView) view.findViewById(R.id.bottom_navigation_small_item_icon);
+				final TextView notification = (TextView) view.findViewById(R.id.bottom_navigation_notification);
 
 				icon.setSelected(false);
-				AHHelper.updateTopMargin(icon, activeMarginTop, inactiveMargin);
-				AHHelper.updateLeftMargin(notification, notificationActiveMarginLeft, notificationInactiveMarginLeft);
-				AHHelper.updateTextColor(title, itemActiveColor, itemInactiveColor);
+
+				if (titleState != TitleState.ALWAYS_HIDE) {
+					AHHelper.updateTopMargin(icon, activeMarginTop, inactiveMargin);
+					AHHelper.updateLeftMargin(notification, notificationActiveMarginLeft, notificationInactiveMarginLeft);
+					AHHelper.updateTopMargin(notification, notificationActiveMarginTop, notificationInactiveMarginTop);
+					AHHelper.updateTextColor(title, itemActiveColor, itemInactiveColor);
+					AHHelper.updateWidth(container, selectedItemWidth, notSelectedItemWidth);
+				}
+
 				AHHelper.updateAlpha(title, 1, 0);
-				AHHelper.updateWidth(container, selectedItemWidth, notSelectedItemWidth);
 				AHHelper.updateDrawableColor(context, items.get(currentItem).getDrawable(context), icon,
 						itemActiveColor, itemInactiveColor, forceTint);
-
 			}
 		}
 
@@ -665,12 +838,12 @@ public class AHBottomNavigation extends FrameLayout {
 		if (currentItem > 0 && currentItem < items.size()) {
 			currentColor = items.get(currentItem).getColor(context);
 		} else if (currentItem == CURRENT_ITEM_NONE) {
-			setBackgroundColor(defaultBackgroundColor);
+			if (defaultBackgroundResource != 0) {
+				setBackgroundResource(defaultBackgroundResource);
+			} else {
+				setBackgroundColor(defaultBackgroundColor);
+			}
 			backgroundColorView.setBackgroundColor(Color.TRANSPARENT);
-		}
-
-		if (tabSelectedListener != null && useCallback) {
-			tabSelectedListener.onTabSelected(itemIndex, false);
 		}
 	}
 
@@ -685,13 +858,17 @@ public class AHBottomNavigation extends FrameLayout {
 				continue;
 			}
 
+			final AHNotification notificationItem = notifications.get(i);
+			final int currentTextColor = AHNotificationHelper.getTextColor(notificationItem, notificationTextColor);
+			final int currentBackgroundColor = AHNotificationHelper.getBackgroundColor(notificationItem, notificationBackgroundColor);
+
 			TextView notification = (TextView) views.get(i).findViewById(R.id.bottom_navigation_notification);
 
 			String currentValue = notification.getText().toString();
-			boolean animate = !currentValue.equals(String.valueOf(notifications[i]));
+			boolean animate = !currentValue.equals(String.valueOf(notificationItem.getText()));
 
 			if (updateStyle) {
-				notification.setTextColor(notificationTextColor);
+				notification.setTextColor(currentTextColor);
 				if (notificationTypeface != null) {
 					notification.setTypeface(notificationTypeface);
 				} else {
@@ -706,19 +883,19 @@ public class AHBottomNavigation extends FrameLayout {
 						notification.setBackgroundDrawable(notificationBackgroundDrawable);
 					}
 
-				} else if (notificationBackgroundColor != 0) {
+				} else if (currentBackgroundColor != 0) {
 					Drawable defautlDrawable = ContextCompat.getDrawable(context, R.drawable.notification_background);
 					if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.JELLY_BEAN) {
 						notification.setBackground(AHHelper.getTintDrawable(defautlDrawable,
-								notificationBackgroundColor, forceTint));
+								currentBackgroundColor, forceTint));
 					} else {
 						notification.setBackgroundDrawable(AHHelper.getTintDrawable(defautlDrawable,
-								notificationBackgroundColor, forceTint));
+								currentBackgroundColor, forceTint));
 					}
 				}
 			}
 
-			if (notifications[i].length() == 0 && notification.getText().length() > 0) {
+			if (notificationItem.isEmpty() && notification.getText().length() > 0) {
 				notification.setText("");
 				if (animate) {
 					notification.animate()
@@ -729,8 +906,8 @@ public class AHBottomNavigation extends FrameLayout {
 							.setDuration(150)
 							.start();
 				}
-			} else if (notifications[i].length() > 0) {
-				notification.setText(String.valueOf(notifications[i]));
+			} else if (!notificationItem.isEmpty()) {
+				notification.setText(String.valueOf(notificationItem.getText()));
 				if (animate) {
 					notification.setScaleX(0);
 					notification.setScaleY(0);
@@ -844,6 +1021,16 @@ public class AHBottomNavigation extends FrameLayout {
 	}
 
 	/**
+	 * Set the bottom navigation background resource
+	 *
+	 * @param defaultBackgroundResource The bottom navigation background resource
+	 */
+	public void setDefaultBackgroundResource(@DrawableRes int defaultBackgroundResource) {
+		this.defaultBackgroundResource = defaultBackgroundResource;
+		createItems();
+	}
+
+	/**
 	 * Get the accent color (used when the view contains 3 items)
 	 *
 	 * @return The default accent color
@@ -884,6 +1071,26 @@ public class AHBottomNavigation extends FrameLayout {
 	}
 
 	/**
+	 * Set the colors used when the bottom bar uses the colored mode
+	 *
+	 * @param colorActive   The active color
+	 * @param colorInactive The inactive color
+	 */
+	public void setColoredModeColors(@ColorInt int colorActive, @ColorInt int colorInactive) {
+		this.coloredTitleColorActive = colorActive;
+		this.coloredTitleColorInactive = colorInactive;
+		createItems();
+	}
+
+	/**
+	 * Set selected background visibility
+     */
+	public void setSelectedBackgroundVisible(boolean visible) {
+		this.selectedBackgroundVisible = visible;
+		createItems();
+	}
+
+	/**
 	 * Set notification typeface
 	 *
 	 * @param typeface Typeface
@@ -894,7 +1101,7 @@ public class AHBottomNavigation extends FrameLayout {
 	}
 
 	/**
-	 * Set title text size
+	 * Set title text size in pixels
 	 *
 	 * @param activeSize
 	 * @param inactiveSize
@@ -902,6 +1109,18 @@ public class AHBottomNavigation extends FrameLayout {
 	public void setTitleTextSize(float activeSize, float inactiveSize) {
 		this.titleActiveTextSize = activeSize;
 		this.titleInactiveTextSize = inactiveSize;
+		createItems();
+	}
+
+	/**
+	 * Set title text size in SP
+	 *
+	 +	 * @param activeSize in sp
+	 +	 * @param inactiveSize in sp
+	 */
+	public void setTitleTextSizeInSp(float activeSize, float inactiveSize) {
+		this.titleActiveTextSize = TypedValue.applyDimension(TypedValue.COMPLEX_UNIT_SP, activeSize, resources.getDisplayMetrics());
+		this.titleInactiveTextSize = TypedValue.applyDimension(TypedValue.COMPLEX_UNIT_SP, inactiveSize, resources.getDisplayMetrics());
 		createItems();
 	}
 
@@ -948,7 +1167,8 @@ public class AHBottomNavigation extends FrameLayout {
 			return;
 		}
 
-		if (items.size() == MIN_ITEMS || forceTitlesDisplay) {
+		if (titleState != TitleState.ALWAYS_HIDE &&
+				(items.size() == MIN_ITEMS || titleState == TitleState.ALWAYS_SHOW)) {
 			updateItems(position, useCallback);
 		} else {
 			updateSmallItems(position, useCallback);
@@ -974,15 +1194,31 @@ public class AHBottomNavigation extends FrameLayout {
 		if (getParent() instanceof CoordinatorLayout) {
 			ViewGroup.LayoutParams params = getLayoutParams();
 			if (bottomNavigationBehavior == null) {
-				bottomNavigationBehavior = new AHBottomNavigationBehavior<>(behaviorTranslationEnabled);
+				bottomNavigationBehavior = new AHBottomNavigationBehavior<>(behaviorTranslationEnabled, navigationBarHeight);
 			} else {
-				bottomNavigationBehavior.setBehaviorTranslationEnabled(behaviorTranslationEnabled);
+				bottomNavigationBehavior.setBehaviorTranslationEnabled(behaviorTranslationEnabled, navigationBarHeight);
+			}
+			if (navigationPositionListener != null) {
+				bottomNavigationBehavior.setOnNavigationPositionListener(navigationPositionListener);
 			}
 			((CoordinatorLayout.LayoutParams) params).setBehavior(bottomNavigationBehavior);
 			if (needHideBottomNavigation) {
 				needHideBottomNavigation = false;
 				bottomNavigationBehavior.hideView(this, bottomNavigationHeight, hideBottomNavigationWithAnimation);
+				isHidden = true;
 			}
+		}
+	}
+
+	/**
+	 * Manage the floating action button behavior with AHBottomNavigation
+	 * @param fab Floating Action Button
+	 */
+	public void manageFloatingActionButtonBehavior(FloatingActionButton fab) {
+		if (fab.getParent() instanceof CoordinatorLayout) {
+			AHBottomNavigationFABBehavior fabBehavior = new AHBottomNavigationFABBehavior(navigationBarHeight);
+			((CoordinatorLayout.LayoutParams) fab.getLayoutParams())
+					.setBehavior(fabBehavior);
 		}
 	}
 
@@ -999,6 +1235,7 @@ public class AHBottomNavigation extends FrameLayout {
 	 * @param withAnimation Boolean
 	 */
 	public void hideBottomNavigation(boolean withAnimation) {
+		isHidden = true;
 		if (bottomNavigationBehavior != null) {
 			bottomNavigationBehavior.hideView(this, bottomNavigationHeight, withAnimation);
 		} else if (getParent() instanceof CoordinatorLayout) {
@@ -1027,6 +1264,7 @@ public class AHBottomNavigation extends FrameLayout {
 	 * @param withAnimation Boolean
 	 */
 	public void restoreBottomNavigation(boolean withAnimation) {
+		isHidden = false;
 		if (bottomNavigationBehavior != null) {
 			bottomNavigationBehavior.resetOffset(this, withAnimation);
 		} else {
@@ -1037,6 +1275,20 @@ public class AHBottomNavigation extends FrameLayout {
 					.setDuration(withAnimation ? 300 : 0)
 					.start();
 		}
+	}
+
+	/**
+	 * Return if the translucent navigation is enabled
+	 */
+	public boolean isTranslucentNavigationEnabled() {
+		return translucentNavigationEnabled;
+	}
+
+	/**
+	 * Set the translucent navigation value
+	 */
+	public void setTranslucentNavigationEnabled(boolean translucentNavigationEnabled) {
+		this.translucentNavigationEnabled = translucentNavigationEnabled;
 	}
 
 	/**
@@ -1060,22 +1312,25 @@ public class AHBottomNavigation extends FrameLayout {
 	}
 
 	/**
-	 * Return if we force the titles to be displayed
+	 * Return the title state for display
 	 *
-	 * @return Boolean
+	 * @return TitleState
 	 */
-	public boolean isForceTitlesDisplay() {
-		return forceTitlesDisplay;
+	public TitleState getTitleState() {
+		return titleState;
 	}
 
 	/**
-	 * Force the titles to be displayed (or used the classic behavior)
-	 * Note: Against Material Design guidelines
+	 * Sets the title state for each tab
+	 * SHOW_WHEN_ACTIVE: when a tab is focused
+	 * ALWAYS_SHOW: show regardless of which tab is in focus
+	 * ALWAYS_HIDE: never show tab titles
+	 * Note: Always showing the title is against Material Design guidelines
 	 *
-	 * @param forceTitlesDisplay Boolean
+	 * @param titleState TitleState
 	 */
-	public void setForceTitlesDisplay(boolean forceTitlesDisplay) {
-		this.forceTitlesDisplay = forceTitlesDisplay;
+	public void setTitleState(TitleState titleState) {
+		this.titleState = titleState;
 		createItems();
 	}
 
@@ -1094,6 +1349,26 @@ public class AHBottomNavigation extends FrameLayout {
 	}
 
 	/**
+	 * Set OnNavigationPositionListener
+	 */
+	public void setOnNavigationPositionListener(OnNavigationPositionListener navigationPositionListener) {
+		this.navigationPositionListener = navigationPositionListener;
+		if (bottomNavigationBehavior != null) {
+			bottomNavigationBehavior.setOnNavigationPositionListener(navigationPositionListener);
+		}
+	}
+
+	/**
+	 * Remove OnNavigationPositionListener()
+	 */
+	public void removeOnNavigationPositionListener() {
+		this.navigationPositionListener = null;
+		if (bottomNavigationBehavior != null) {
+			bottomNavigationBehavior.removeOnNavigationPositionListener();
+		}
+	}
+
+	/**
 	 * Set the notification number
 	 *
 	 * @param nbNotification int
@@ -1102,22 +1377,42 @@ public class AHBottomNavigation extends FrameLayout {
 	@Deprecated
 	public void setNotification(int nbNotification, int itemPosition) {
 		if (itemPosition < 0 || itemPosition > items.size() - 1) {
-			Log.w(TAG, "The position is out of bounds of the items (" + items.size() + " elements)");
-			return;
+            throw new IndexOutOfBoundsException(String.format(Locale.US, EXCEPTION_INDEX_OUT_OF_BOUNDS, itemPosition, items.size()));
 		}
-		notifications[itemPosition] = nbNotification == 0 ? "" : String.valueOf(nbNotification);
+        final String title = nbNotification == 0 ? "" : String.valueOf(nbNotification);
+        notifications.set(itemPosition, AHNotification.justText(title));
 		updateNotifications(false, itemPosition);
 	}
 
 	/**
-	 * Set Notification content
+	 * Set notification text
 	 *
-	 * @param title String
+	 * @param title        String
 	 * @param itemPosition int
 	 */
 	public void setNotification(String title, int itemPosition) {
-		notifications[itemPosition] = title;
-		updateNotifications(false, itemPosition);
+        if (itemPosition < 0 || itemPosition > items.size() - 1) {
+            throw new IndexOutOfBoundsException(String.format(Locale.US, EXCEPTION_INDEX_OUT_OF_BOUNDS, itemPosition, items.size()));
+        }
+        notifications.set(itemPosition, AHNotification.justText(title));
+        updateNotifications(false, itemPosition);
+    }
+
+    /**
+     * Set fully customized Notification
+     *
+     * @param notification AHNotification
+     * @param itemPosition int
+     */
+	public void setNotification(AHNotification notification, int itemPosition) {
+		if (itemPosition < 0 || itemPosition > items.size() - 1) {
+            throw new IndexOutOfBoundsException(String.format(Locale.US, EXCEPTION_INDEX_OUT_OF_BOUNDS, itemPosition, items.size()));
+		}
+        if (notification == null) {
+            notification = new AHNotification(); // instead of null, use empty notification
+        }
+		notifications.set(itemPosition, notification);
+		updateNotifications(true, itemPosition);
 	}
 
 	/**
@@ -1182,6 +1477,7 @@ public class AHBottomNavigation extends FrameLayout {
 
 	/**
 	 * Set the notification margin left
+	 *
 	 * @param activeMargin
 	 * @param inactiveMargin
 	 */
@@ -1204,12 +1500,33 @@ public class AHBottomNavigation extends FrameLayout {
 
 	/**
 	 * Activate or not the elevation, and set the value
+	 *
 	 * @param useElevation boolean
-	 * @param elevation float
+	 * @param elevation    float
 	 */
 	public void setUseElevation(boolean useElevation, float elevation) {
 		ViewCompat.setElevation(this, useElevation ? elevation : 0);
 		setClipToPadding(false);
+	}
+
+	/**
+	 * Return if the Bottom Navigation is hidden or not
+	 */
+	public boolean isHidden() {
+		return isHidden;
+	}
+
+	/**
+	 * Get the view at the given position
+	 * @param position int
+	 * @return The view at the position, or null
+	 */
+	public View getViewAtPosition(int position) {
+		if (linearLayoutContainer != null && position >= 0
+				&& position < linearLayoutContainer.getChildCount()) {
+			return linearLayoutContainer.getChildAt(position);
+		}
+		return null;
 	}
 
 	////////////////
@@ -1220,7 +1537,23 @@ public class AHBottomNavigation extends FrameLayout {
 	 *
 	 */
 	public interface OnTabSelectedListener {
-		void onTabSelected(int position, boolean wasSelected);
+		/**
+		 * Called when a tab has been selected (clicked)
+		 *
+		 * @param position    int: Position of the selected tab
+		 * @param wasSelected boolean: true if the tab was already selected
+		 * @return boolean: true for updating the tab UI, false otherwise
+		 */
+		boolean onTabSelected(int position, boolean wasSelected);
+	}
+
+	public interface OnNavigationPositionListener {
+		/**
+		 * Called when the bottom navigation position is changed
+		 *
+		 * @param y int: y translation of bottom navigation
+		 */
+		void onPositionChange(int y);
 	}
 
 }