+++ demo/src/main/java/com/aurelhubert/ahbottomnavigation/demo/DemoActivity.java
@@ -1,7 +1,8 @@
 package com.aurelhubert.ahbottomnavigation.demo;
 
 import android.animation.Animator;
-import android.graphics.Color;
+import android.content.Context;
+import android.content.Intent;
 import android.os.Bundle;
 import android.os.Handler;
 import android.support.design.widget.FloatingActionButton;
@@ -13,8 +14,10 @@ import android.view.View;
 import android.view.animation.OvershootInterpolator;
 
 import com.aurelhubert.ahbottomnavigation.AHBottomNavigation;
+import com.aurelhubert.ahbottomnavigation.AHBottomNavigationAdapter;
 import com.aurelhubert.ahbottomnavigation.AHBottomNavigationItem;
 import com.aurelhubert.ahbottomnavigation.AHBottomNavigationViewPager;
+import com.aurelhubert.ahbottomnavigation.notification.AHNotification;
 
 import java.util.ArrayList;
 
@@ -22,7 +25,11 @@ public class DemoActivity extends AppCompatActivity {
 
 	private DemoFragment currentFragment;
 	private DemoViewPagerAdapter adapter;
+	private AHBottomNavigationAdapter navigationAdapter;
 	private ArrayList<AHBottomNavigationItem> bottomNavigationItems = new ArrayList<>();
+	private boolean useMenuResource = true;
+	private int[] tabColors;
+	private Handler handler = new Handler();
 
 	// UI
 	private AHBottomNavigationViewPager viewPager;
@@ -32,8 +39,17 @@ public class DemoActivity extends AppCompatActivity {
 	@Override
 	protected void onCreate(Bundle savedInstanceState) {
 		super.onCreate(savedInstanceState);
+		boolean enabledTranslucentNavigation = getSharedPreferences("shared", Context.MODE_PRIVATE)
+				.getBoolean("translucentNavigation", false);
+		setTheme(enabledTranslucentNavigation ? R.style.AppTheme_TranslucentNavigation : R.style.AppTheme);
 		setContentView(R.layout.activity_home);
 		initUI();
+	}
+
+	@Override
+	protected void onDestroy() {
+		super.onDestroy();
+		handler.removeCallbacksAndMessages(null);
 	}
 
 	/**
@@ -45,25 +61,36 @@ public class DemoActivity extends AppCompatActivity {
 		viewPager = (AHBottomNavigationViewPager) findViewById(R.id.view_pager);
 		floatingActionButton = (FloatingActionButton) findViewById(R.id.floating_action_button);
 
-		AHBottomNavigationItem item1 = new AHBottomNavigationItem(R.string.tab_1, R.drawable.ic_apps_black_24dp, R.color.color_tab_1);
-		AHBottomNavigationItem item2 = new AHBottomNavigationItem(R.string.tab_2, R.drawable.ic_maps_local_bar, R.color.color_tab_2);
-		AHBottomNavigationItem item3 = new AHBottomNavigationItem(R.string.tab_3, R.drawable.ic_maps_local_restaurant, R.color.color_tab_3);
+		if (useMenuResource) {
+			tabColors = getApplicationContext().getResources().getIntArray(R.array.tab_colors);
+			navigationAdapter = new AHBottomNavigationAdapter(this, R.menu.bottom_navigation_menu_3);
+			navigationAdapter.setupWithBottomNavigation(bottomNavigation, tabColors);
+		} else {
+			AHBottomNavigationItem item1 = new AHBottomNavigationItem(R.string.tab_1, R.drawable.ic_apps_black_24dp, R.color.color_tab_1);
+			AHBottomNavigationItem item2 = new AHBottomNavigationItem(R.string.tab_2, R.drawable.ic_maps_local_bar, R.color.color_tab_2);
+			AHBottomNavigationItem item3 = new AHBottomNavigationItem(R.string.tab_3, R.drawable.ic_maps_local_restaurant, R.color.color_tab_3);
 
-		bottomNavigationItems.add(item1);
-		bottomNavigationItems.add(item2);
-		bottomNavigationItems.add(item3);
+			bottomNavigationItems.add(item1);
+			bottomNavigationItems.add(item2);
+			bottomNavigationItems.add(item3);
 
-		bottomNavigation.addItems(bottomNavigationItems);
-		bottomNavigation.setAccentColor(Color.parseColor("#F63D2B"));
-		bottomNavigation.setInactiveColor(Color.parseColor("#747474"));
+			bottomNavigation.addItems(bottomNavigationItems);
+		}
+
+		bottomNavigation.manageFloatingActionButtonBehavior(floatingActionButton);
+		bottomNavigation.setTranslucentNavigationEnabled(true);
 
 		bottomNavigation.setOnTabSelectedListener(new AHBottomNavigation.OnTabSelectedListener() {
 			@Override
-			public void onTabSelected(int position, boolean wasSelected) {
-;
+			public boolean onTabSelected(int position, boolean wasSelected) {
+
+				if (currentFragment == null) {
+					currentFragment = adapter.getCurrentFragment();
+				}
+
 				if (wasSelected) {
 					currentFragment.refresh();
-					return;
+					return true;
 				}
 
 				if (currentFragment != null) {
@@ -144,8 +171,18 @@ public class DemoActivity extends AppCompatActivity {
 								.start();
 					}
 				}
+
+				return true;
 			}
 		});
+
+		/*
+		bottomNavigation.setOnNavigationPositionListener(new AHBottomNavigation.OnNavigationPositionListener() {
+			@Override public void onPositionChange(int y) {
+				Log.d("DemoActivity", "BottomNavigation Position: " + y);
+			}
+		});
+		*/
 
 		viewPager.setOffscreenPageLimit(4);
 		adapter = new DemoViewPagerAdapter(getSupportFragmentManager());
@@ -153,15 +190,23 @@ public class DemoActivity extends AppCompatActivity {
 
 		currentFragment = adapter.getCurrentFragment();
 
-		final Handler handler = new Handler();
 		handler.postDelayed(new Runnable() {
 			@Override
 			public void run() {
-				bottomNavigation.setNotification("16", 1);
+				// Setting custom colors for notification
+				AHNotification notification = new AHNotification.Builder()
+						.setText(":)")
+						.setBackgroundColor(ContextCompat.getColor(DemoActivity.this, R.color.color_notification_back))
+						.setTextColor(ContextCompat.getColor(DemoActivity.this, R.color.color_notification_text))
+						.build();
+				bottomNavigation.setNotification(notification, 1);
 				Snackbar.make(bottomNavigation, "Snackbar with bottom navigation",
 						Snackbar.LENGTH_SHORT).show();
+
 			}
 		}, 3000);
+
+		//bottomNavigation.setDefaultBackgroundResource(R.drawable.bottom_navigation_background);
 	}
 
 	/**
@@ -183,20 +228,32 @@ public class DemoActivity extends AppCompatActivity {
 	 */
 	public void updateBottomNavigationItems(boolean addItems) {
 
-		AHBottomNavigationItem item4 = new AHBottomNavigationItem(getString(R.string.tab_4),
-				ContextCompat.getDrawable(this, R.drawable.ic_maps_local_bar),
-				ContextCompat.getColor(this, R.color.color_tab_4));
-		AHBottomNavigationItem item5 = new AHBottomNavigationItem(getString(R.string.tab_5),
-				ContextCompat.getDrawable(this, R.drawable.ic_maps_place),
-				ContextCompat.getColor(this, R.color.color_tab_5));
+		if (useMenuResource) {
+			if (addItems) {
+				navigationAdapter = new AHBottomNavigationAdapter(this, R.menu.bottom_navigation_menu_5);
+				navigationAdapter.setupWithBottomNavigation(bottomNavigation, tabColors);
+				bottomNavigation.setNotification("1", 3);
+			} else {
+				navigationAdapter = new AHBottomNavigationAdapter(this, R.menu.bottom_navigation_menu_3);
+				navigationAdapter.setupWithBottomNavigation(bottomNavigation, tabColors);
+			}
 
-		if (addItems) {
-			bottomNavigation.addItem(item4);
-			bottomNavigation.addItem(item5);
-			bottomNavigation.setNotification("1", 3);
 		} else {
-			bottomNavigation.removeAllItems();
-			bottomNavigation.addItems(bottomNavigationItems);
+			if (addItems) {
+				AHBottomNavigationItem item4 = new AHBottomNavigationItem(getString(R.string.tab_4),
+						ContextCompat.getDrawable(this, R.drawable.ic_maps_local_bar),
+						ContextCompat.getColor(this, R.color.color_tab_4));
+				AHBottomNavigationItem item5 = new AHBottomNavigationItem(getString(R.string.tab_5),
+						ContextCompat.getDrawable(this, R.drawable.ic_maps_place),
+						ContextCompat.getColor(this, R.color.color_tab_5));
+
+				bottomNavigation.addItem(item4);
+				bottomNavigation.addItem(item5);
+				bottomNavigation.setNotification("1", 3);
+			} else {
+				bottomNavigation.removeAllItems();
+				bottomNavigation.addItems(bottomNavigationItems);
+			}
 		}
 	}
 
@@ -209,6 +266,29 @@ public class DemoActivity extends AppCompatActivity {
 		} else {
 			bottomNavigation.hideBottomNavigation(true);
 		}
+	}
+
+	/**
+	 * Show or hide selected item background
+	 */
+	public void updateSelectedBackgroundVisibility(boolean isVisible) {
+		bottomNavigation.setSelectedBackgroundVisible(isVisible);
+	}
+
+	/**
+	 * Show or hide selected item background
+	 */
+	public void setForceTitleHide(boolean forceTitleHide) {
+		AHBottomNavigation.TitleState state = forceTitleHide ? AHBottomNavigation.TitleState.ALWAYS_HIDE : AHBottomNavigation.TitleState.ALWAYS_SHOW;
+		bottomNavigation.setTitleState(state);
+	}
+
+	/**
+	 * Reload activity
+	 */
+	public void reload() {
+		startActivity(new Intent(this, DemoActivity.class));
+		finish();
 	}
 
 	/**